Extract tag for attachment upload
We handle file uploads in several places, and we seem to have some
common patterns around them:
- an input element
- optionally a link to download the existing, previously uploaded
  attachment
- a hidden element containing the previously uploaded attachment ID

In order to better handle the misleading "No file selected" message,
first extract the common functionality from each of the different forms
that have an attachment upload form. This will allow us to make sure our
forms are consistent in how they handle files, and allow us to later
change the behavior of all the forms at once.

Use the new tag in an individual provider's license information step,
and in an organizational provider's facility license step.

Issue #158 "No file selected" is confusing when a file is uploaded