Write wrapper method for enrollment notifications: #770
Write a method to add the submitter (the user who submitted the
application) to all notifications about changes to the status of an
enrollment.  Use this method to send the approval notification.  Next, I
will add it to the pending and modified notifications.

Adding it to rejection will happen after #750 is merged.