Use new wrapper to send other notifications: #770
Wrap modified and submitted notifications with the new wrapper.  Also,
remove an unnecessary variable in AcceptedHandler.

In the wrapper, use contact name instead of "submittedBy."  The
enrollment may be edited and resubmitted by an administrator.  In that
case, the email would be addressed to them when it should be addressed
to the original creator of the application.  Since we use contact email
to get the email address, the contact name is the best way to address
the recipient.