Show and log exception in Spring handler
We have a global exception handler which shows a very generic message.
This is appropriate for a production system, but we're not yet a
production system, and it is frustrating to have such a noninformative
error message.

Show the text of the exception in the web view, and log the full
exception and the relevant method + URL (but not query parameters, both
for potential privacy reasons and because it is distractingly difficult
to present them well).